backend/requirements.txt
flask
google-cloud-speech
moviepy